New game in-game rework
Fixed some stuff that made generating new game using the * button in-game a little weird. The window would just freeze leaving the user wandering what in the world is going on. Now when the user presses the button a neat overlay pops up informing the user that generation is on underway. A timer is there to show the user the app isn't frozen. And if you hold shift while pressing the button the generation is animated. Which is...neat, I-I guess?