Network training and estimation
This is like a huge commit with a ton of finished and unfinished stuff. The training part works for the most part. Like there are definitely some mistakes, but the majority works. The second part, consisting of adaptations to GridCore and creation of DrawableGridCore, which is like the base of drawing a grid without the controls, is not finished. The drawing itself __works__. It draws. Really. Slowly. But it draws. I need to find where the ticks leak and fix it. Oh and also networks do mistakes. So further training is needed. Now it gets to a point where it just can't place any more numbers and gets stuck. And it fucks up and crashes on heap space. So that...