Drawable rework
Drawable is no longer a PGraphics2D subclass, instead all the methods were reimplemented and use an internal PGraphics, or more optimally the canvas of the applet. This way it might be just a little more effective. I hope... There are still some minor issues, the controls and the clock show up weird.