Small fixes, removed //image(...) commented-out calls, Java formatting
Buttons in Notwork Training are where they are supposed to, someone might have forgotten to put a rectMode(CORNER) at the beginning of the Button's draw function and someone might have added some other rectMode to the view.
All the image calls were removed as they were remnants of the old times of Drawable, when each one was a separate PGraphics. That is no longer the case and all the Drawables actually render when the update function is called.
I wanted to see how the C# formatting would look and I really liked it, but then got sick of it. It looks good in VS, howeved IDEA just makes it look like shit. So that is gone now. Yaaay. I guess. Blame IntelliJ.