Scoreboard serialization and storage
The Scoreboard had some issues in the way it was serialized as a java object. If I changed anything anywhere the version id of the class changed and I had to set it manually to fix the issue. I feel like that is not the cleanest way so it is now saved as a zip file of a CSV and its hash. The hash is there to prevent users from manually adding good scores or removing bad ones. I mean, you can still do it but you'd have to get the hash in the same way the program does, which I guess is not that difficult, but anyways...
Also something similar will have to be done with the neural network part as there are still the version ids.